test(swingset): test that baggage and durables survive vat upgrade
We create a durable Kind, and reattach behavior to it in v2. The
handle must travel through baggage, demonstrating that baggage works.

I'm still looking for the right way to use VatData these from with
swingset tests.. other packages should import @agoric/vat-data, but
that might be circular from here

refs #1848